[REFACTOR] implement OOP and some of SOLID principle
